+++ R/model_spatial_day.R
@@ -13,6 +13,15 @@
     met = meteoland::interpolationpoints(meteo, spt, dates=as.Date(date), verbose=FALSE)
     met = met@data[[1]]
   }
+  else if(inherits(meteo, "stars")) {
+    pt_sf = sf::st_sf(geometry = xi$point, elevation = xi$elevation, slope = xi$slope, aspect = xi$aspect)
+    met = meteoland::interpolate_data(pt_sf, meteo, dates = as.Date(date))
+    met = met$interpolated_data[[1]]
+    met = as.data.frame(met)
+    row.names(met) = met$dates
+    met$dates = NULL
+  }    
+  
   tmin = met[date,"MinTemperature"]
   tmax = met[date,"MaxTemperature"]
   rhmin = met[date,"MinRelativeHumidity"]