+++ tests/testthat/test_spatial.R
@@ -6,7 +6,6 @@ ypts = sp_to_sf(examplepointslandscape)
 yws = sp_to_sf(examplewatershed)
 
 data("examplemeteo")
-data("exampleinterpolationdata")
 data("SpParamsMED")
 dates = seq(as.Date("2001-03-01"), as.Date("2001-03-03"), by="day")
 
@@ -22,6 +21,14 @@ test_that("Can simulate three days over landscape",{
 })
 
 # test_that("Can simulate three days over landscape using old interpolator",{
-#   expect_s3_class(spwb_spatial(ypts[1,], meteo = exampleinterpolationdata, dates = dates, 
+#   data("exampleinterpolationdata")
+#   expect_s3_class(spwb_spatial(ypts[1,], meteo = exampleinterpolationdata, 
 #                                SpParams = SpParamsMED, progress = FALSE), "sf")
 # })
+
+test_that("Can simulate three days over landscape using new interpolator",{
+  interpolator = with_meteo(meteoland_meteo_example) |>
+    create_meteo_interpolator(params = defaultInterpolationParams())
+  expect_s3_class(spwb_spatial(ypts[1,], meteo = interpolator,
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+})