+++ tests/testthat/test_spatial_day.R
@@ -21,6 +21,14 @@ test_that("Can simulate one day over landscape",{
 })
 # 
 # test_that("Can simulate one day over landscape using old interpolator",{
+#   data("exampleinterpolationdata")
 #   expect_s3_class(spwb_spatial_day(ypts[1,], meteo = exampleinterpolationdata, date = date, 
 #                                    SpParams = SpParamsMED, progress = FALSE), "sf")
+# })
+
+# test_that("Can simulate three days over landscape using new interpolator",{
+#   interpolator = with_meteo(meteoland_meteo_example) |>
+#     create_meteo_interpolator(params = defaultInterpolationParams())
+#     expect_s3_class(spwb_spatial_day(ypts[1,], meteo = interpolator, date = "2022-04-01", 
+#                                      SpParams = SpParamsMED, progress = FALSE), "sf")
 # })